remove duplicated stuff and reorganize
